Updated git submodules
Project: yangtools master 501f8d250a4dec41b5b7ebb77693878726616025

BUG-7052: remove AbstractEffectiveSchemaContext

This is essentially a copy of AbstractSchemaContext, so reduce
code duplication by removing it.

Change-Id: I736017a379c4796a7afed6933525b354f1f852bc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>


BUG-6972: Do not allow root StmtContext to be copied

It cannot be targeted, hence it does not make sense to copy it.
Removes a chunk of duplicate code and takes this class out of the
picture where copy operations are concerned.

Change-Id: I7a8d46ba59658a534534cd8204c43d894d017247
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>