Updated git submodules
Project: aaa master 8c31b022c29326adc9a05000d24d28e72b8bc919

Merge "Bug 7793: shiro.ini needs to die"

Bug 7793: shiro.ini needs to die

shiro.ini is a poor means for configuration in our configuration-rich
ODL context.  Some primary concerns involve the fact that shiro.ini
is not replicated nor consistent across a cluster of ODL nodes.
Prior to this change, lack of proper SOA means that services couldn't
be dynamically decided at runtime (i.e., swap out backing
implementation).

Simply put, shiro.ini sucks.  This was a known limitation when shiro
was first integrated, but was swallowed hook line and sinker for the
other advantages that the integration provided.  It is time for us to
change this so that shiro is configured using a more mature and robust
mechanism (i.e., clustered-app-config).

Some known limitations in this patch include the fact that, because we
utilize jersey 1.X with its static configuration through web.xml, there
is a race between when the configuration is available, and when the
web service is instantiated.  To overcome this limitation, a basic
waiting mechanism was added to KarafIniWebEnvironment to wait until
the configuration is available prior to initializing Shiro.

The name "aaa-app-config" was chosen on purpose-- this is due to the
fact that most AAA related properties should eventually be configured
through this single pane solution.  This will reduce complexity and
reliance on running disparate ManagedService instances.

Change-Id: Ie87886dd57958957dfd0a642fec434862903b509
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>