Updated git submodules
Project: yangtools master 15ae10fcef8efa8ed397c9cac9be596e92a39eec

Bug 8745: Add support for parsing attributes to the new XML parser

New XML parser is now able to parse XML attributes in order to provide
necessary information for processing edit-config messages in NETCONF.

Make some members of ImmutableNormalizedNodeStreamWriter
protected so they can be reused
in the EditOperationNormalizedNodeStreamWriter extension in NETCONF.

Add a unit test

This patch is a prerequisite for the following change in NETCONF:
https://git.opendaylight.org/gerrit/#/c/60014/

Change-Id: Idfd87279814e324dc914809fb09abfbb27eefa4a
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>


Bug 8715: Add strictParsing switch to the new XML parser

Add the option to switch strict parsing mode on/off to
the new XML parser. If the mode is switched on, the parser
will throw an exception if it encounters unknown child nodes
in YANG containers and lists. If the mode is off, the parser
will skip such nodes.

Change-Id: Ieece9b15ced06282a43877c2c1e884515217244b
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>