Create simple bash script to generate rel-notes
The release notes generator Java tool does not seem to be working for
the Carbon release. As we do not have the expertise to fix it this is an
attempt to mimic the tool using a simple bash script.

Change-Id: I18277c5e4e36a5db1ed619953680f6201eadfc6d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>
(cherry picked from commit 994fa9efb58b18bd71caf246c9ab64c2dae484ef)