Updated git submodules
Project: openflowplugin stable/nitrogen 97ddaf6a08fee6218cd1aa88d0aa11d6f5357b1e

Bug 8902 - Reconciliation Framework changes

Changes related to the Reconciliation Framework are done in this patch.
The APIs needed to be implemented by the services registering to the
Reconciliation framework are present in the interface
ReconciliationNotificationListener.The APIs to register the service to
Reconciliation framework and the APIs to start the reconciliation /
cancellation of the ongoing reconciliation of a node are present in
ReconciliationManager.

Change-Id: I8bd1845db8b86f4a4314977e05c6b699592f5658
Signed-off-by: gobinath <gobinath@ericsson.com>