Updated git submodules
Project: aaa master fad3387edd1bb5e05f2ef244f16fca398b606f92

Fix hang in KarafIniWebEnvironment#init

It's possible the call to getInstance races with the BP container startup
in which case the INSTANCE will be created with all nulls and thus the
shiroConfiguration will forever be null. To get a valid INSTANCE, it must
be created via BP which injects a non-null ShiroConfiguration.

So getInstance now just returns INSTANCE and KarafIniWebEnvironment#init
busy waits on it.

Change-Id: I9c22d0dbec580c15830ece10c7dfa27f68ab8acf
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>