Updated git submodules
Project: yangtools master f8da8a59e37cb3b668b759c86b3d680a1a6aad71

Add XmlParserStream.traverse(DOMSource)

This adds an alternative traversal based on DOMSource. This is needed
as it seems XMLStreamReaders based on DOMSource are in short supply and
in some cases we are interpreting a part of an already-parsed Document
as a NormalizedNode.

Change-Id: I3e21daa144ca90d868f71b3f5af318f4c35cc24f
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>