Updated git submodules
Project: of-config master 6efb06bd1a8d4f9172f7322403a2c470ac29ed10

Convert to blueprint and remove CSS remnants

Convert the project to use blueprint as the config subsystem is
being deprecated in Nitrogen.

The ofconf-nodes-config.xml file was also removed because the
mechanism of pushing netcong connector configs via the CSS
controller-config loopback has been removed in Nitrogen. I'm
not sure if this config is legacy and not used/needed anymore
or needs to be replaced with something else.

Change-Id: I0547f684b19bdfc582d9cd41961f08c18eebc7cd
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>